Move the gulp tasks into a subdirectory.
