+++ gulp/tasks/watch.js
@@ -2,7 +2,7 @@
 // which lists what tasks are available.
 var gulp = require('gulp-help')(require('gulp'));
 
-var projectPaths = require('../package.json')['paths'];
+var projectPaths = require('../../package.json')['paths'];
 var serverGlobs = [].concat(projectPaths['server-js'], projectPaths['view-templates']);
 
 gulp.task('watch:start', 'Start watching files for changes.', function(done) {
